+++ sdmx-dl-cli/src/test/java/sdmxdl/cli/ListFlowsCommandTest.java
@@ -43,8 +43,8 @@ public class ListFlowsCommandTest {
                 .isEmpty();
 
         assertThat(FileSample.readAll(out))
-                .contains("Ref,Label", atIndex(0))
-                .contains("all:data&struct(latest),data", atIndex(1))
+                .contains("Ref,Label,Description", atIndex(0))
+                .contains("all:data&struct(latest),data,", atIndex(1))
                 .hasSize(2);
     }
 }